security_helper/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

daiquiri==3.2.5.1
netifaces==0.11.0 ; sys_platform == 'linux'
psutil==6.1.0
pycryptodomex==3.21.0

# custom
pigar==2.1.6
pyflakes==3.2.0
